server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black[d]>=25.12.0",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-auth>=2.45.0",
    "pydantic>=2.12.5",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
